Created model.js
Model.js is what the interface for a type revolves around, in this case Des3d. 
The parameters are enclosed between {}s, which are put under attributes, which are put under definitions{} 
- One can customize the parameters' appearances, with "ui:" or size. There are ways not listed here, such as conditional parameters, which can be found in mesh_conditional. Unfortunately, I could not find out how to stick them into the "mesh" view/attribute.
- Perhaps it was not necessary to put size = 1 in each parameter

At the very top are views, which attributes are assigned to. I don't understand fully how it works, but attributes go under views (the latter being side tabs), creating more side tabs. 
- As said in the Simput documentation, one can nest many attributes/parameters (confused) inside each other
- The order of the views is dictated by "order" at the very top.

One can assign hooks, I think to attributes (the attributes can have 1 parameter each), which allow Simput to be more dynamic. Hooks are not included in this project/code, but one example is in the Simput documentation for the Oscillator model. One can also look into more complex models, such as Pyfr or Vera. 

Des3d requires an input file with a certain format, which requires a template file. Simput can still return the parameters of a data model, however, as a JSON file.